requirements-all.txt
scikit-learn==0.19.1
pandas==0.23.4
passlib==1.7.1
gensim==3.7.1
numpy==1.16.2
tqdm==4.31.1
networkx==2.4
tensorflow-gpu==1.15.0
keras-bert==0.80.0
keras-adaptive-softmax==0.6.0
